requirements.txt
streamlit==1.31.1
ultralytics==8.1.2
opencv-python==4.9.0.80
pytesseract==0.3.10
pillow==10.2.0
pandas==2.2.0
numpy==1.26.3
streamlit-webrtc==0.47.1
av==10.0.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
torch==2.1.0
torchvision==0.16.0
scikit-image==0.22.0
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
pathlib==1.0.1
tqdm==4.66.1
PyYAML==6.0.1